communication_scenario/requirements.txt
concrete-python==2.10.0
filelock==3.18.0
fsspec==2025.5.0
importlib_resources==6.5.2
Jinja2==3.1.6
joblib==1.5.0
jsonpickle==4.1.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mpmath==1.3.0
networkx==3.4.2
numpy==1.26.4
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
pycryptodome==3.23.0
Pygments==2.19.1
rich==14.0.0
scikit-learn==1.6.1
scipy==1.15.3
setuptools==80.8.0
sympy==1.14.0
threadpoolctl==3.6.0
torch==2.7.0
triton==3.3.0
typing_extensions==4.13.2
wheel==0.45.1
z3-solver==4.13.0.0
